tradeoffers: GetTradeOffers: add TradeFilterItemDescriptions
This adds a parameter to steam web api to get item descriptions for items
Note: This does not automatically parse them, it gives the user control over it
so that it does not add extra fields to EconItem.

Also added EconDesc to InventoryItem instead to reduce boilerplate.

I've updated the tradeoffers example on how to parse it.

Signed-off-by: A. Samy <f.fallen45@gmail.com>